src_code_analysis/requirements.txt
openai
pymupdf
databricks-sdk
pandas
pygments
radon
bandit
streamlit-ace
streamlit